Toppings

+++ src/main/scala/example/Order.scala
@@ -2,5 +2,5 @@ package example
 
 final case class Order(
   pizzaCount: Int,
-  value: Int
+  value: BigDecimal
 )
